Moves test sql fixtures into the fixtures class
