+++ tests/table_repository_test.py
@@ -1,5 +1,6 @@
 from unittest import TestCase
 
+from tests.test_fixtures import TestFixtures
 from schema.repository.table_repository import TableRepository
 from schema.factory.table_factory import TableFactory
 from schema.models.table import Table
@@ -9,9 +10,10 @@ class TableRepositoryTest(TestCase):
         super().setUp()
         factory = TableFactory()
         self._repo = TableRepository(table_factory=factory)
+        self._fixtures = TestFixtures()
 
     def test_get_table(self) -> None:
-        sql = self._get_test_sql()
+        sql = self._fixtures._get_test_table_sql()
 
         self._repo.load_from_string(source=sql)
 
@@ -20,39 +22,10 @@ class TableRepositoryTest(TestCase):
         self.assertIsInstance(result, Table)
 
     def test_get_table_for_missing_name(self) -> None:
-        sql = self._get_test_sql()
+        sql = self._fixtures._get_test_table_sql()
 
         self._repo.load_from_string(source=sql)
 
         result = self._repo.get_table('missing')
 
         self.assertIsNone(result)
-
-    def _get_test_sql(self) -> str:
-
-        return """
-        CREATE TABLE people
-  (
-     id       INT UNIQUE NOT NULL,
-     EMAIL     NVARCHAR(400),
-     PRIMARY KEY(ID),
-     FAMILY_ID INT
-  )
-    CREATE TABLE car
-  (
-     id       INT PRIMARY KEY,
-     NAME     NVARCHAR(400),
-     OWNER_ID INT,
-     CONSTRAINT fk_category FOREIGN KEY (OWNER_ID)
-                         REFERENCES people(ID)
-  )
-
-      CREATE TABLE FAMILY
-  (
-     id       INT PRIMARY KEY,
-     NAME     NVARCHAR(400)
-  )
-
-    ALTER TABLE people ADD FOREIGN KEY (FAMILY_ID) REFERENCES FAMILY(ID)
-  ;
-        """
