+++ tests/test_fixtures.py
@@ -38,3 +38,32 @@ class TestFixtures:
     "constraints": {},
     "tablespace": ""
 }
+
+    def _get_test_table_sql(self) -> str:
+
+        return """
+        CREATE TABLE people
+  (
+     id       INT UNIQUE NOT NULL,
+     EMAIL     NVARCHAR(400),
+     PRIMARY KEY(ID),
+     FAMILY_ID INT
+  )
+    CREATE TABLE car
+  (
+     id       INT PRIMARY KEY,
+     NAME     NVARCHAR(400),
+     OWNER_ID INT,
+     CONSTRAINT fk_category FOREIGN KEY (OWNER_ID)
+                         REFERENCES people(ID)
+  )
+
+      CREATE TABLE FAMILY
+  (
+     id       INT PRIMARY KEY,
+     NAME     NVARCHAR(400)
+  )
+
+    ALTER TABLE people ADD FOREIGN KEY (FAMILY_ID) REFERENCES FAMILY(ID)
+  ;
+        """
